Add files via upload
In this notebook, I describe the process of finding the cross spectral density between two audio signals. The cross-spectral density is the Fourier transform of the cross-correlation between the two signals. Two signals (normalized) which are alike each other should exhibit a greater cross-correlation  than two unlike signals. From this code, we see that ball-racket impact signals seem to exhibit a high frequency peak in the cross-spectral density that is absent in that between a ball-racket signal and what we consider to be random noise. Files needed: CorrelationsOfAudioSignals.ipynb, BF48.wav, BF28.wav, BF85.wav, and BallRacketWithBabolatAtHome.wav